fix(drawio): narrow the asset route and the MathJax suppression to when they apply
Two review findings.

The asset route answered on every deployment, including those left on draw.io's own hosted
viewer, where nothing requests it: the client rebases the viewer's asset paths onto the
route only for a self-hosted instance, and viewer.diagrams.net sends the CORS header that
makes the route unnecessary in the first place. It now checks the same condition the client
does -- isSelfHostedDrawio, moved up so both halves share one definition -- and 404s
otherwise, so a default deployment carries no reachable outbound fetch.

adoptMathJax() left window.MathJax holding a placeholder object when the baked-in location
could not be read, which broke math outright (initMath() had been suppressed and was never
re-run) and left a global behind that anything treating `typeof window.MathJax !==
'undefined'` as "MathJax is present" would trip over. It now always clears the placeholder
and re-runs initMath(), asking for draw.io's own location when there is nothing to relocate
onto -- so that path ends up exactly as it would have been without any of this.

The README records why the global is safe to touch at all: GROWI's own page math is
remark-math + rehype-katex, KaTeX never reads window.MathJax, and viewer-static.min.js
already sets it on every page view. It also records the two deployment choices that remove
the need for the route -- draw.io behind a reverse proxy on GROWI's origin, or an instance
configured to send Access-Control-Allow-Origin.
